Rename components names and add confirmation modal
- Minor renames
- Add confirmation modal
- Fix typo
